syscall: Make io_uring syscall arguments consistent
Make the arguments of io_uring syscalls consistent with what is said in
the manpage. No functional change is intended.

Link: https://lore.kernel.org/io-uring/CADUfDZpE_gPyfN=dLKB6nu-++ZKyebpWTvYGNOmdP1-c_BLZZA@mail.gmail.com
Link: https://lore.kernel.org/io-uring/CADUfDZr0mPn_REb24aEPa477T+CYeoV5hcbURqX9kazCUqRp4A@mail.gmail.com
Suggested-by: Caleb Sander <csander@purestorage.com>
Link: https://lore.kernel.org/r/20220830005122.885209-2-ammar.faizi@intel.com

+++ src/arch/generic/syscall.h
@@ -5,15 +5,15 @@
 
 #include <fcntl.h>
 
-static inline int __sys_io_uring_register(int fd, unsigned opcode,
-					  const void *arg, unsigned nr_args)
+static inline int __sys_io_uring_register(unsigned int fd, unsigned int opcode,
+					  const void *arg, unsigned int nr_args)
 {
 	int ret;
 	ret = syscall(__NR_io_uring_register, fd, opcode, arg, nr_args);
 	return (ret < 0) ? -errno : ret;
 }
 
-static inline int __sys_io_uring_setup(unsigned entries,
+static inline int __sys_io_uring_setup(unsigned int entries,
 				       struct io_uring_params *p)
 {
 	int ret;
@@ -21,9 +21,10 @@ static inline int __sys_io_uring_setup(unsigned entries,
 	return (ret < 0) ? -errno : ret;
 }
 
-static inline int __sys_io_uring_enter2(int fd, unsigned to_submit,
-					unsigned min_complete, unsigned flags,
-					sigset_t *sig, int sz)
+static inline int __sys_io_uring_enter2(unsigned int fd, unsigned int to_submit,
+					unsigned int min_complete,
+					unsigned int flags, sigset_t *sig,
+					size_t sz)
 {
 	int ret;
 	ret = syscall(__NR_io_uring_enter, fd, to_submit, min_complete, flags,
@@ -31,9 +32,9 @@ static inline int __sys_io_uring_enter2(int fd, unsigned to_submit,
 	return (ret < 0) ? -errno : ret;
 }
 
-static inline int __sys_io_uring_enter(int fd, unsigned to_submit,
-				       unsigned min_complete, unsigned flags,
-				       sigset_t *sig)
+static inline int __sys_io_uring_enter(unsigned int fd, unsigned int to_submit,
+				       unsigned int min_complete,
+				       unsigned int flags, sigset_t *sig)
 {
 	return __sys_io_uring_enter2(fd, to_submit, min_complete, flags, sig,
 				     _NSIG / 8);
